requirements.txt
python-telegram-bot==20.7
aiohttp>=3.8.5
nest_asyncio